requirements.txt
line-bot-sdk==2.0.1
Flask==2.0.3
python-dotenv==0.20.0
